Removed the server-side Python/Celery-backed game simulation code, leaving only the client-side JavaScript version.
Why, you ask? Well, the original plan was to keep both in parallel. This
would pay dividends down the rode if someone wanted to make a
multiplayer version of the game, where you wouldn't want critical code
running client-side where users could easily modify it. However, the
game is currently single player and will likely remain so for a while.
And as I added the JavaScript version, I realized that it was going to
need a quite different infrastructure to support it than the Python
version. So I deleted game_sim.py, leaving only game_sim.js. If anyone
wants to add back server-side game simulation at some point in the
future, it wouldn't be very difficult.