Changes to Probabilities for Extreme Heights
There are a few changes to the height probability distribution.  The
main change is that players 7'8"+ are now less common - about every 50
draft classes (every 7 or 8 new leagues) instead of every 20-30 draft
classes (every 3 or 4 new leagues).

Added 3andD and RawAthlete cases to the if-else statement in
genRatings().  They aren't being used, but I put them there for
completeness in case there's a reason to specify those profiles in the
future.  (Currently the only place a profile is specified is the Lavar
Ball easter egg.)