Update genMessage.js
Hey, this is user lilpaki from reddit. I made a dozen or so changes. If you like my work, I'll continue to add in my freetime. Feel free to only use the ones you like. I did make a few meme-y ones that might not be to your taste. Thanks!