Moved from a 2 layer Tanh (with bias) to a 4 layer ReLU (no bias).
Why? ReLU layers without bias are input scale-invariant (they're just
multiplies and thresholds at zero). So this doesn't care if the ratings
are near 20 or near 80... it can't tell the difference.

I needed the extra layers to get comparable performance. Although it is
slightly worse in terms of loss, I think being input scale invariant is a huge win.